Make default choices of parameters from user defined fields
This adds functions in thermos-specs.magic-fields, which when applied
to a document will amend the document so settings like civil costs &
tariffs are chosen for candidates based on certain user-defined fields.

The distinction between this and a manual edit is that the manual edit
relates by ID and is done in the UI.

However we also want to be able to load in prepared data for which we
know what civils etc. we want to use. This relation has to be done on
the basis of name rather than ID, because IDs are hidden. It also
needs to take effect whenever a candidate has been loaded or created
or when the name of a blob of settings is changed.

Because of this, the join is done at the end, before using the
document, by passing it through magic-fields/join (which see). This is
a bit horrible, since it has to be done before rendering the document
to the UI, since if you click on something you want to see what it's
using.

Similarly it needs to be done before running the problem, so the right
effects happen.

However it shouldn't be done before save, as it is not "fixing" the
choices - they can be overridden by later changes.

ALso, we special case the user-defined field for CONNECTION-COST:

The connection cost is handled specially - if it's a number, that
number is taken as the connection cost.