Ensure that we do not have any bits of old solution lying around
With addition of magic fields, it is possible to have bits of old
solution stored on candidates that didn't get included in the problem
sent to server (because those candidates are not marked to keep, since
nobody touched them). Since the solver strips solution only from what
its sent, and that is merged back into the state, we can get dangling
results from last solution until we reload.

This way purges the solution data from any ephemeral candidates, so
that can't happen.