Add a field to alternatives to model storage in house
This is a neater version of what we did for Bristol Zero Carbon.

The THERMOS peak load value represents the peak heat load coming out
of stuff in the building, so includes things like instant hot water.
In a heat network this is fine, because the network can deliver the
30kW you need to run your shower.

For some types of individual system, especially heat pumps, it makes
more sense to use a hot water tank to reduce the peak required out of
the machine. Then the HP can be sized for the space heating peak which
is smaller.

In this change I've added a field to alternatives which is called
kwp-per-mean-kw. If this contains a value, then all the per-kWp costs
are shifted into the kWh costs for the optimiser according to this
ratio / 8760 (so the network model is unaware of this feature)

In the UI this is called 'Tank factor'.

This change also updates the UI to display the revised peak, where appropriate.